Catch and log errors raised by container.kill()
In the finally block, if the container can't be killed (because
it's already died with an error, say) then container.remove()
would not get called. This meant that the docker daemon would get
cluttered with more and more broken containers from failed test
runs.

This change makes it so that docker APIErrors are caught and logged
but do not block further cleanup.